+++ src/cli/tc.rs
@@ -16,7 +16,7 @@ use crate::cli::server::Server;
 use crate::cli::{self, CLIError};
 use crate::core::blueprint::Blueprint;
 use crate::core::config::reader::ConfigReader;
-use crate::core::generator::Reader;
+use crate::core::generator::ConfigConsoleGenerator;
 use crate::core::http::API_URL_PREFIX;
 use crate::core::print_schema;
 use crate::core::rest::{EndpointSet, Unchecked};
@@ -82,7 +82,9 @@ pub async fn run() -> Result<()> {
         }
         Command::Init { folder_path } => init(&folder_path).await,
         Command::Gen { file_path } => {
-            Reader::new(runtime.clone(), &file_path).generate().await?;
+            ConfigConsoleGenerator::new(&file_path, runtime.clone())
+                .generate()
+                .await?;
 
             Ok(())
         }
